Refactor: Change code structure
