+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from 'react';
 import { Route, Routes } from 'react-router';
-import Layout from './components/layout/Layout';
+import Layout from './components/Layout/Layout';
 import AddTask from './pages/AddTask';
 import EditTask from './pages/EditTask';
 import MonthCalendar from './pages/MonthCalendar';
