+++ src/components/Navigation/Navigation.styled.tsx
@@ -1,0 +1,17 @@
+import styled from '@emotion/styled';
+
+export const NavWrapper = styled.nav`
+  background: #faf9f9;
+`;
+
+export const Ul = styled.ul`
+  height: 100%;
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+
+  li > a {
+    display: block;
+    padding: 20px;
+  }
+`;
