+++ src/components/Task/Task.styled.tsx
@@ -1,0 +1,5 @@
+import styled from '@emotion/styled';
+
+export const TaskWrapper = styled.div`
+  display: flex;
+`;
