+++ src/components/layout/Layout.types.ts
@@ -1,0 +1,3 @@
+export type LayoutProps = {
+  children: React.ReactNode;
+};
